server/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.3
flask-restx==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
Werkzeug==2.2.3
